requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile -U
#
anyio==4.0.0
    # via httpcore
beautifulsoup4==4.12.2
    # via bs4
bs4==0.0.1
    # via -r requirements.in
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.2.0
    # via requests
h11==0.14.0
    # via httpcore
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via python-telegram-bot
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
lxml==4.9.3
    # via -r requirements.in
python-telegram-bot==20.5
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
soupsieve==2.5
    # via beautifulsoup4
urllib3==2.0.5
    # via requests
